MYR-447 Split the purge target table out of purge.go
Adding eight label targets pushed purge.go to 571 lines. The declarative
table and the engine that walks it are cleanly separable, so Targets (and
its rationale) moves to targets.go: 450 + 126, both under the cap.

No behaviour change.
